pyproject.toml
# File: pyproject.toml
# Description: Project metadata and dependencies (uv format)

[project]
name = "aitube"
version = "0.1.0"
description = "Next-generation video platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    "celery[redis]>=5.3.6",
    "python-multipart>=0.0.9",
    "orjson>=3.9.14",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "polars>=0.20.10",
    "minio>=7.2.4",
    "ffmpeg-python>=0.2.0",
    "watchdog>=4.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
strict = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

[tool.hatch.build.targets.wheel]
packages = ["src"]